+++ komp-ai/src/app/page.tsx
@@ -4,6 +4,8 @@ import DnaSection from "../components/home/DnaSection";
 import KursySection from "../components/home/KursySection";
 import HumanMachineSection from "../components/home/HumanMachineSection";
 import ProwadzacySection from "../components/home/ProwadzacySection";
+import KontaktSection from "../components/home/KontaktSection";
+import StopkaSection from "../components/home/StopkaSection";
 import { getHomePage, getSiteSettings } from "../content";
 
 export default async function Home() {
@@ -20,6 +22,8 @@ export default async function Home() {
       <KursySection cards={page.kursTiles} />
       <HumanMachineSection hm={page.humanMachine} />
       <ProwadzacySection p={page.prowadzacy} />
+      <KontaktSection k={page.kontakt} />
+      <StopkaSection f={settings.footer} />
     </main>
   );
 }